Cleanup index.js, move getHeaderBytes
