+++ src/util/getHeaderBytes.js
